% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_Date.R
\name{date_to_Date}
\alias{date_to_Date}
\title{Convert dates of most standard formats to class \code{Date} (base R)}
\usage{
date_to_Date(date)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}
}
\value{
Returns entered date or datetime as date of class \code{Date}
(base R) with time zone "America/Los_Angeles".
}
\description{
Dates and datetimes may be written in many formats and may be of many
different data classes in R. \code{date_to_Date()} converts dates and
datetimes of most standard formats into base R's \code{Date} class. Date and
datetime formats not accepted are any containing words like "January" or
abbreviations like "January". NOTE: \code{date_to_Date()} is run internally
by all exported waterCALendar functions that take \code{date} as an argument.
}
\examples{
date_to_Date("2024-11-01 12:00:00")
date_to_Date(as.POSIXct("2024-11-01"))
date_to_Date("11/01/2024 12:00:00")
date_to_Date("11/1/2024")

}
\seealso{
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}}
\code{\link[=date_to_wyt]{date_to_wyt()}}
}
