% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_dowy.R
\name{date_to_dowy}
\alias{date_to_dowy}
\title{Convert Date to Day of Water Year (DOWY)}
\usage{
date_to_dowy(date, verbose = FALSE)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}

\item{verbose}{Logical. Display reminder about years vs water years?
Default is \code{FALSE}.}
}
\value{
Returns a \code{numeric} day of water year (DOWY).
}
\description{
California water years begin on October 1. \code{date_to_dowy()} returns the
day of water year when you enter a date.
}
\details{
Note that if the date entered falls between October 1 and
December 31 (inclusive), the DOWY
returned will be DOWY of the year following that of the date entered.
If \code{verbose=TRUE}, a
reminder message to this effect will be displayed.
}
\examples{
date_to_dowy("2024-01-01",verbose=T)
date_to_dowy(as.Date("2024-01-01"))
date_to_dowy(as.POSIXct("2024-01-01"))
date_to_dowy("1/1/2024 12:00")
date_to_dowy(as.POSIXlt("2024-01-01"),verbose=T)

}
\seealso{
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=date_to_wy]{date_to_wy()}},
\code{\link[=date_to_wyt]{date_to_wyt()}},
\code{\link[=date_to_Date]{date_to_Date()}}
}
