% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_doy.R
\name{date_to_doy}
\alias{date_to_doy}
\title{Convert Date to Day of Year (DOY)}
\usage{
date_to_doy(date, verbose = FALSE)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}

\item{verbose}{Logical. Display explanatory messages with output? Default is
\code{FALSE}.}
}
\value{
Returns a \code{numeric} day of year (DOY).
}
\description{
\code{date_to_doy()} returns the DOY of the date entered.
}
\details{
Note that the DOY returned begins as 1 on January 1, whereas
\code{as.POSIXlt("yyyy-mm-dd")$yday} begins as 0 on January 1.
}
\examples{
date_to_doy("2024-01-01")
date_to_doy(as.Date("2024-01-01"), verbose=TRUE)
date_to_doy("1/1/2024")
date_to_doy(as.POSIXct("2024-01-01 12:45:30"), verbose=TRUE)
date_to_doy(as.POSIXlt("2024-01-01 12:45:30"))

}
\seealso{
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_wy]{date_to_wy()}},
\code{\link[=date_to_wyt]{date_to_wyt()}},
\code{\link[=date_to_Date]{date_to_Date()}}
}
