% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_mowy.R
\name{date_to_mowy}
\alias{date_to_mowy}
\title{Convert Date to Month of Water Year (MOWY)}
\usage{
date_to_mowy(date)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}
}
\value{
Returns a \code{numeric} month of water year (MOWY) between 1 and 12
(inclusive).
}
\description{
California water years begin on October 1. \code{\link[=date_to_mowy]{date_to_mowy()}} returns the month
of water year (MOWY). For example, if the date is in December, the MOWY is 3.
}
\examples{
date_to_mowy("2024-01-01")
date_to_mowy(as.Date("2024-01-01"))
date_to_mowy("1/1/2024")
date_to_mowy(as.POSIXct("2024-01-01"))
date_to_mowy(as.POSIXlt("2024-01-01"))

}
\seealso{
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_wy]{date_to_wy()}},
\code{\link[=date_to_wyt]{date_to_wyt()}},
\code{\link[=date_to_Date]{date_to_Date()}}
}
