% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_wy.R
\name{date_to_wy}
\alias{date_to_wy}
\title{Convert Date to Water Year}
\usage{
date_to_wy(date, verbose = FALSE)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}

\item{verbose}{Logical. Display reminder about years vs water years? Default
is \code{FALSE}.}
}
\value{
Returns \code{numeric} water year of the date entered.
}
\description{
California water years begin on October 1. \code{date_to_wy()} returns the
water year when you enter a date. In practice, this means that if the date is
before October 1, water year is just year, otherwise water year is year+1.
}
\examples{
date_to_wy("2024-11-01")
date_to_wy("1/1/2024")
date_to_wy(as.Date("2024-11-01"))
date_to_wy(as.POSIXct("2024-11-01 12:45:30"))

}
\seealso{
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}}
\code{\link[=date_to_wyt]{date_to_wyt()}},
\code{\link[=date_to_Date]{date_to_Date()}}
}
