% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_wyt.R
\name{date_to_wyt}
\alias{date_to_wyt}
\title{Convert Date to Water Year Type}
\usage{
date_to_wyt(date, region, abbreviate = F)
}
\arguments{
\item{date}{A date or datetime of any format EXCEPT those containing words
like "January" or abbreviations like "Jan". Accepted classes are
\code{character}, \code{Date}, \code{POSIXct} and \code{POSIXlt}.}

\item{region}{Sacramento or San Joaquin Valley? Region may be abbreviated
(or not) in any of several ways, e.g., "San", "sj", "San Joaquin", "sac",
"Sacramento", etc. Only \code{character} is accepted.}

\item{abbreviate}{Abbreviate the water year type (e.g., show "Dry" as "D")?}
}
\value{
Returns \code{character} water year type of the date and region
entered.
}
\description{
California water years (beginning October 1) in the Sacramento and San
Joaquin Valleys are separately assigned one of five hydrologic
classifications (water year types, WYT), based on measured unimpaired runoff:
Wet (W); Above Normal (AN); Below Normal (BN); Dry (D); or Critical (C).
\code{date_to_wyt()} returns the water year type when you enter a date and a
region (Sacramento or San Joaquin).
}
\examples{
date_to_wyt("2024-11-01", "Sac)
date_to_wyt(as.Date("2024-11-01"), "SJ)
date_to_wyt("1/1/2024","San Joaquin")

}
\seealso{
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}}
\code{\link[=date_to_Date]{date_to_Date()}}
}
