% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowy_to_date.R
\name{dowy_to_date}
\alias{dowy_to_date}
\title{Convert Day of Water Year (DOWY) to Date}
\usage{
dowy_to_date(dowy, wyear)
}
\arguments{
\item{dowy}{A day of water year from 1 to 365; 366 is accepted in leap years.
Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{wyear}{A four-digit water year. Accepted classes are \code{numeric}, \code{character} and \code{factor}.}
}
\value{
Returns the date of the \code{dowy} entered as a \code{Date} object.
}
\description{
California water years begin on October 1. \code{dowy_to_date()} returns the
date of the \code{dowy} entered. Water year (\code{wyear}) is required, to
account for leap years.
}
\examples{
dowy_to_date(45,"1999")
dowy_to_date("45",1999)
dowy_to_date("45","1999",verbose=F)

}
\seealso{
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
