% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dowy_to_doy.R
\name{dowy_to_doy}
\alias{dowy_to_doy}
\title{Convert Day of Water Year (DOWY) to day of year (DOY)}
\usage{
dowy_to_doy(dowy, wyear, verbose = FALSE)
}
\arguments{
\item{dowy}{A day of water year between 1 and 365 (inclusive)' 366 accepted in leap years.
Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{wyear}{A four-digit water year. Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{verbose}{Logical. Display reminder about years vs water years? Default is \code{FALSE}.}
}
\value{
Returns a \code{numeric} day of year (DOY). If \code{dowy} <= 92 (i.e., if the date is
October 1 - December 31), the DOY returned will be DOY of the year preceding the (\code{wyear}) entered.
}
\description{
California water years begin on October 1. \code{dowy_to_doy()} returns the
DOY if you enter a DOWY and a year.
}
\details{
If \code{dowy} <= 92, the DOY returned is DOY of the year preceding
the water year (\code{wyear}) entered. If \code{verbose=TRUE}, a reminder message
to this effect will be displayed.
}
\examples{
dowy(1,"1999")
dowy("365",1999)
dowy("150","1999",verbose=T)

}
\seealso{
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
