% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_to_date.R
\name{doy_to_date}
\alias{doy_to_date}
\title{Convert Day of Year (DOY) to Date}
\usage{
doy_to_date(doy, year)
}
\arguments{
\item{doy}{A day of year between 1 and 365 (inclusive); 366 accepted in leap years.
Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{year}{A four-digit year. Accepted classes are \code{numeric}, \code{character} and \code{factor}.}
}
\value{
Returns the date (as a \code{Date} object) of the \code{doy} entered.
}
\description{
\code{doy_to_date()} returns the DOY of the date entered. \code{year} is required
to account for leap years.
}
\details{
Note that the \code{doy} argument assumes that \code{doy} 1 is January 1,
whereas \code{as.POSIXlt("yyyy-mm-dd")$yday} begins as 0 on January 1.
}
\examples{
doy_to_date("10",2022)
doy_to_date(10,"2022")
doy_to_date("10","2022")

}
\seealso{
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
