% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_to_dowy.R
\name{doy_to_dowy}
\alias{doy_to_dowy}
\title{Convert Day of Year (DOY) to Day of Water Year (DOWY)}
\usage{
doy_to_dowy(doy, year, verbose = FALSE)
}
\arguments{
\item{doy}{A day of year, entered as a number from 1 to 365 (inclusive); 366 is accepted in leap years.
Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{year}{A four-digit year. Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{verbose}{Logical. Display reminder about years vs water years? Default is \code{FALSE}.}
}
\value{
Returns a \code{numeric} day of water year (DOWY), where water year begins on 01 October.
}
\description{
California water years begin on October 1. \code{doy_to_dowy()} returns the
day of the year (DOWY) if you enter a day of year.
}
\details{
Note that the \code{doy} argument assumes that \code{doy} 1 is January 1,
whereas \code{as.POSIXlt("yyyy-mm-dd")$yday} begins as 0 on January 1.
}
\examples{
doy_to_dowy(150,"2024")
doy_to_dowy("150",2024,verbose=TRUE)
doy_to_dowy("150","2024")

}
\seealso{
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
