% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_to_mowy.R
\name{month_to_mowy}
\alias{month_to_mowy}
\title{Convert Month to Month of Water Year (MOWY)}
\usage{
month_to_mowy(month)
}
\arguments{
\item{month}{A month, entered as a number from 1 to 12 (inclusive), a three-letter month abbreviation,
(e.g. "Jan") or a full month name (e.g. "January"). Accepted classes are \code{numeric}, \code{character} and \code{factor}.}
}
\value{
Returns a \code{numeric} month of water year (MOWY).
}
\description{
California water years begin on October 1. \code{month_to_mowy()} returns the
month of water year (MOWY) if you enter a month.
}
\examples{
month_to_mowy(5,"1999")
month_to_mowy("Mar",1999)
month_to_mowy("December","1999")

}
\seealso{
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
