% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mowy_to_month.R
\name{mowy_to_month}
\alias{mowy_to_month}
\title{Convert Month of Water Year (MOWY) to Month}
\usage{
mowy_to_month(mowy, form = "numeric")
}
\arguments{
\item{mowy}{A month of water year (MOWY) between 1 and 12 (inclusive).
Accepted classes are \code{numeric}, \code{character} and \code{factor}.}

\item{form}{One of "numeric", "Mmm" or "long". Default is "numeric".}
}
\value{
If form = "numeric", returns a \code{numeric} month of the year, 1 (January) to 12 (December).
If form = "Mmm" returns a three-letter month code (e.g. "Jan") as an ordered \code{factor}. If form = "long",
returns a full month name (e.g. "January") as an ordered \code{factor}.
}
\description{
California water years begin on October 1. \code{mowy_to_month()} returns the
month of the year if you enter a month of water year (MOWY).
}
\examples{
mowy_to_month(5,form="Mmm")
mowy_to_month("5",form="long")
mowy_to_month("5")

}
\seealso{
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_wy]{date_to_wy()}}
}
