% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pkg_env}
\alias{pkg_env}
\title{Run function \code{update_wyt} whenever \code{library(waterCALendar)} is
called. Not exported to namespace:waterCALendar.
Create package environment to hold updated data}
\format{
An object of class \code{environment} of length 0.
}
\usage{
pkg_env
}
\description{
Run function \code{update_wyt} whenever \code{library(waterCALendar)} is
called. Not exported to namespace:waterCALendar.
Create package environment to hold updated data
}
\keyword{datasets}
