% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_wyi.R
\name{uniform_wyi}
\alias{uniform_wyi}
\title{Convert water year type abbreviations to full names, e.g., "D" to "Dry". This
function is not exported to namespace:waterCALendar, but is run internally by
\code{update_wyt()} (also not exported).}
\usage{
uniform_wyi(dat)
}
\arguments{
\item{dat}{\code{character} column or vector containing a water year type
abbreviation, as in: "W", "AN", "BN", "D" or "C".}
}
\description{
Convert water year type abbreviations to full names, e.g., "D" to "Dry". This
function is not exported to namespace:waterCALendar, but is run internally by
\code{update_wyt()} (also not exported).
}
