% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wyt.R
\name{update_wyt}
\alias{update_wyt}
\title{If waterCALendar::wy_indices dataset is not current, \code{update_wyt()}
imports recent water year index and water year type data from
"https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST". This function
is not exported to namespace:waterCALendar, but is run whenever
\code{library(waterCALendar)} is called. \code{update_wyt()} will return a
warning if no internet connection is found.}
\usage{
update_wyt()
}
\description{
If waterCALendar::wy_indices dataset is not current, \code{update_wyt()}
imports recent water year index and water year type data from
"https://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST". This function
is not exported to namespace:waterCALendar, but is run whenever
\code{library(waterCALendar)} is called. \code{update_wyt()} will return a
warning if no internet connection is found.
}
