% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterCALendar-package.R
\docType{package}
\name{waterCALendar-package}
\alias{waterCALendar}
\alias{waterCALendar-package}
\title{waterCALendar: Calendar Conversions for the California Water Year Calendar}
\description{
Tired of trying to calculate day of water year (DOWY) from the date? Or day of year (DOY) from DOWY? Or month of water year (MOWY) from the date? This package contains functions to do all these things and much more. The function names should be self-explanatory, but feel free to email me if you have any questions, or if the package doesn't do something (relevant) you'd really like it to do. Most importantly, please let me know if any of these functions doesn't work the way it's supposed to or the way you want it to. Aesthetic changes by popular demand only. Functions contained in this package include doy_to_dowy(), dowy_to_doy(), date_to_dowy(), date_to_doy(), dowy_to_date(), doy_to_date(), date_to_mowy(), month_to_mowy(), mowy_to_month(), date_to_wy(), date_to_wyt(), date_to_Date(). Datasets included: wy_indices (use data(wy_indices) to load).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielrodonnell/waterCALendar}
  \item Report bugs at \url{https://github.com/danielrodonnell/waterCALendar/issues}
}

}
\author{
\strong{Maintainer}: Danny O'Donnell \email{daniel.odonnell@wildlife.ca.gov} (\href{https://orcid.org/0000-0003-0452-2888}{ORCID})

}
\keyword{internal}
