% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wy_to_wyt.R
\name{wy_to_wyt}
\alias{wy_to_wyt}
\title{Convert Date to Water Year Type}
\usage{
wy_to_wyt(wy, region, abbreviate = F)
}
\arguments{
\item{wy}{A water year of the format yyyy. Accepted classes are
\code{character} and \code{numeric}.}

\item{region}{Sacramento or San Joaquin Valley? Region may be abbreviated
(or not) in any of several ways, e.g., "San", "sj", "San Joaquin", "sac",
"Sacramento", etc. Only \code{character} is accepted.}

\item{abbreviate}{Abbreviate the water year type (e.g., show "Dry" as "D")?}
}
\value{
Returns \code{character} water year type of the water year and region
entered.
}
\description{
California water years (beginning October 1) in the Sacramento and San
Joaquin Valleys are separately assigned one of five hydrologic
classifications (water year types, WYT), based on measured unimpaired runoff:
Wet (W); Above Normal (AN); Below Normal (BN); Dry (D); or Critical (C).
\code{date_to_wyt()} returns the water year type when you enter a date and a
region (Sacramento or San Joaquin).
}
\examples{
wy_to_wyt("2024", "Sac", abbreviate=TRUE)
wy_to_wyt(2024, "sj")

}
\seealso{
\code{\link[=date_to_mowy]{date_to_mowy()}},
\code{\link[=mowy_to_month]{mowy_to_month()}},
\code{\link[=month_to_mowy]{month_to_mowy()}},
\code{\link[=date_to_dowy]{date_to_dowy()}},
\code{\link[=dowy_to_date]{dowy_to_date()}},
\code{\link[=doy_to_dowy]{doy_to_dowy()}},
\code{\link[=dowy_to_doy]{dowy_to_doy()}},
\code{\link[=doy_to_date]{doy_to_date()}},
\code{\link[=date_to_doy]{date_to_doy()}},
\code{\link[=date_to_wyt]{date_to_wyt()}},
\code{\link[=date_to_wy]{date_to_wy()}},
\code{\link[=date_to_Date]{date_to_Date()}}
}
